+++ Assets/3Wada/script/PlayreDestroy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd1dc512cc9e56f438629a2324ec73eb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
