+++ Assets/7Konoma/Boss.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 554ad3047e01eed4daf50cc9aca43272
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
