+++ Assets/7Konoma/Images.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7e3751945c84dce4da32e6813dd1222a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
