+++ Assets/7Konoma/prefab/Boss.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &6318185708007150420
+--- !u!1 &436720549379943268
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,13 +8,14 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 6318185708007150416}
-  - component: {fileID: 6318185708007150419}
-  - component: {fileID: 6318185708007150418}
-  - component: {fileID: 6318185708007150415}
-  - component: {fileID: 6318185708007150421}
-  - component: {fileID: 6318185708007150414}
-  - component: {fileID: 6318185708007150417}
+  - component: {fileID: 436720549379943278}
+  - component: {fileID: 436720549379943273}
+  - component: {fileID: 436720549379943272}
+  - component: {fileID: 436720549379943275}
+  - component: {fileID: 436720549379943274}
+  - component: {fileID: 436720549379943269}
+  - component: {fileID: 436720549379943279}
+  - component: {fileID: 436720549379943276}
   m_Layer: 0
   m_Name: Boss
   m_TagString: Untagged
@@ -22,27 +23,27 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &6318185708007150416
+--- !u!4 &436720549379943278
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 11.58, y: -5.3, z: 0}
+  m_LocalPosition: {x: 12, y: -4.5, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!212 &6318185708007150419
+--- !u!212 &436720549379943273
 SpriteRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_Enabled: 1
   m_CastShadows: 0
   m_ReceiveShadows: 0
@@ -76,77 +77,96 @@ SpriteRenderer:
   m_SortingLayerID: 0
   m_SortingLayer: 0
   m_SortingOrder: 0
-  m_Sprite: {fileID: 21300000, guid: 5c2fb025c0892ed48a309f2bb92d97e8, type: 3}
+  m_Sprite: {fileID: 21300000, guid: d26cda33109399d469f87ae0e4c144a2, type: 3}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
   m_FlipX: 0
   m_FlipY: 0
   m_DrawMode: 0
-  m_Size: {x: 1, y: 1}
+  m_Size: {x: 4.047978, y: 2.866}
   m_AdaptiveModeThreshold: 0.5
   m_SpriteTileMode: 0
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!61 &6318185708007150418
-BoxCollider2D:
+--- !u!95 &436720549379943272
+Animator:
+  serializedVersion: 3
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_Enabled: 1
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: 0.0072336197, y: 0.082157135}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1.28, y: 1.28}
-    newSize: {x: 1, y: 1}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 1.371542, y: 1.4445009}
-  m_EdgeRadius: 0
---- !u!61 &6318185708007150415
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: 554ad3047e01eed4daf50cc9aca43272, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorControllerStateOnDisable: 0
+--- !u!61 &436720549379943275
 BoxCollider2D:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
-  m_Offset: {x: 0.020224571, y: 0.0057881474}
+  m_Offset: {x: -0.011092663, y: -0.14723134}
   m_SpriteTilingProperty:
     border: {x: 0, y: 0, z: 0, w: 0}
     pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 1.28, y: 1.28}
-    newSize: {x: 1, y: 1}
+    oldSize: {x: 4.047978, y: 2.866}
+    newSize: {x: 4.047978, y: 2.866}
     adaptiveTilingThreshold: 0.5
     drawMode: 0
     adaptiveTiling: 0
   m_AutoTiling: 0
   serializedVersion: 2
-  m_Size: {x: 1.6361313, y: 1.2915763}
+  m_Size: {x: 4.3559675, y: 2.2207694}
   m_EdgeRadius: 0
---- !u!50 &6318185708007150421
+--- !u!61 &436720549379943274
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 436720549379943268}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 4.047978, y: 2.866}
+    newSize: {x: 4.047978, y: 2.866}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 4, y: 2.5}
+  m_EdgeRadius: 0
+--- !u!50 &436720549379943269
 Rigidbody2D:
   serializedVersion: 4
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_BodyType: 0
   m_Simulated: 1
   m_UseFullKinematicContacts: 0
@@ -160,43 +180,44 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 4
---- !u!114 &6318185708007150414
+--- !u!114 &436720549379943279
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 96b22b6f68df64940aad7a8b158ff936, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _junpHeight: 8
-  _lungesTime: 1
-  _junpCount: 0
-  _rb2: {fileID: 6318185708007150421}
-  _tr: {fileID: 6318185708007150416}
-  _playerTr: {fileID: 0}
-  _speed: 0.01
-  _count:
-  - 900
-  - 500
-  _attackPool: {fileID: 0}
-  _xPos: 0
---- !u!114 &6318185708007150417
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6318185708007150420}
+  m_GameObject: {fileID: 436720549379943268}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: c87b3117517f77f4781a5424727fd3f3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  _hp: 3
-  _continueTime: 1
+  _hp: 2
+  _continueTime: 5
   _jumpForce: 20
   _playerRb2d: {fileID: 0}
+  _stageNumber: 3
+--- !u!114 &436720549379943276
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 436720549379943268}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 96b22b6f68df64940aad7a8b158ff936, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _junpHeight: 3
+  _lungesTime: 2
+  _junpCount: 0
+  _rb2: {fileID: 436720549379943269}
+  _tr: {fileID: 436720549379943278}
+  _playerTr: {fileID: 0}
+  _speed: 0.01
+  _count:
+  - 300
+  - 900
+  _attackPool: {fileID: 0}
+  _xPos: 0
