+++ Assets/7Konoma/prefab/BossAtack.prefab
@@ -11,7 +11,6 @@ GameObject:
   - component: {fileID: 574034840964463355}
   - component: {fileID: 574034840964463348}
   - component: {fileID: 574034840964463349}
-  - component: {fileID: 574034840964463350}
   - component: {fileID: 574034840964463351}
   m_Layer: 0
   m_Name: BossAtack
@@ -72,19 +71,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _speed: 0.1
   _timer: 260
---- !u!114 &574034840964463350
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 574034840964463354}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dd1dc512cc9e56f438629a2324ec73eb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _nowSceneNumber: 0
 --- !u!61 &574034840964463351
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -95,7 +81,7 @@ BoxCollider2D:
   m_Enabled: 1
   m_Density: 1
   m_Material: {fileID: 0}
-  m_IsTrigger: 0
+  m_IsTrigger: 1
   m_UsedByEffector: 0
   m_UsedByComposite: 0
   m_Offset: {x: -1.084732, y: 0.08649027}
@@ -203,6 +189,8 @@ GameObject:
   m_Component:
   - component: {fileID: 574034842223313915}
   - component: {fileID: 574034842223313914}
+  - component: {fileID: 5966023981631902309}
+  - component: {fileID: 8997853738357665024}
   m_Layer: 0
   m_Name: ice
   m_TagString: Untagged
@@ -275,6 +263,45 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &5966023981631902309
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 574034842223313913}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dd1dc512cc9e56f438629a2324ec73eb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _nowSceneNumber: 3
+--- !u!61 &8997853738357665024
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 574034842223313913}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: 0, y: 0}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 1, y: 1}
+    newSize: {x: 1, y: 1}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 1, y: 1}
+  m_EdgeRadius: 0
 --- !u!1 &574034842482601768
 GameObject:
   m_ObjectHideFlags: 0
